pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true
enable_error_code = ["ignore-without-code"]
explicit_package_bases = true
follow_imports = "silent"
no_implicit_reexport = true
plugins = [
  "pydantic.mypy"
]
show_error_codes = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.poetry]
authors = ["Andrés Romero <me@afromero.co>"]
description = ""
name = "pic2panning"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
fonttools = "^4.54.1"
ipython = "^8.29.0"
matplotlib = "^3.9.2"
moviepy = "^1.0.3"
mypy = "^1.13.0"
pillow = "9.5.0" # required for moviepy resize
pre-commit = "^4.0.1"
pydantic = "^2.9.2"
python = "^3.10"
pytubefix = "^8.2.0"
rich = "^13.9.4"
tyro = "^0.8.14"

[tool.poetry.group.generative.dependencies]
accelerate = "^1.1.0"
diffusers = "^0.31.0"
jaxtyping = "^0.2.34"
optimum-quanto = "0.2.2"
protobuf = "^5.28.3"
sam-2 = {git = "git@github.com:facebookresearch/sam2.git"}
sentencepiece = "^0.2.0"
torch = {source = "PytorchCuda", version = "^2.5.1"}
torchvision = {source = "PytorchCuda", version = "^0.20.1"}
transformers = "^4.46.2"

[[tool.poetry.source]]
name = "PytorchCuda"
priority = "explicit"
url = "https://download.pytorch.org/whl/cu121"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
  "poetry.lock"
]
indent-width = 4
# Same as Black.
line-length = 79
# Assume Python 3.10.
target-version = "py310"

[tool.ruff.format]
docstring-code-line-length = "dynamic"
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
# Allow unused variables when underscore-prefixed.
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]

[tool.tomlsort]
all = true
